refactor: customer 관련 dto 패키지 이동
